Extract MCP tool helpers
Move ToolInfo, tool filtering, model-visible schema masking, tool-cache metric naming, and tool qualification into codex-mcp/src/tools.rs.

Update client, manager, apps, exports, and tests to import tool helpers from the new module without behavior changes.
