+++ codex-rs/codex-mcp/src/apps.rs
@@ -3,9 +3,9 @@ use std::path::PathBuf;
 use std::time::Instant;
 
 use crate::mcp::CODEX_APPS_MCP_SERVER_NAME;
-use crate::mcp_connection::MCP_TOOLS_CACHE_WRITE_DURATION_METRIC;
-use crate::mcp_connection::ToolInfo;
 use crate::mcp_connection::emit_duration;
+use crate::tools::MCP_TOOLS_CACHE_WRITE_DURATION_METRIC;
+use crate::tools::ToolInfo;
 use codex_login::CodexAuth;
 use codex_utils_plugins::mcp_connector::is_connector_id_allowed;
 use codex_utils_plugins::mcp_connector::sanitize_name;
