+++ codex-rs/codex-mcp/src/client.rs
@@ -24,13 +24,13 @@ use crate::mcp_connection::DEFAULT_STARTUP_TIMEOUT;
 use crate::mcp_connection::DEFAULT_TOOL_TIMEOUT;
 use crate::mcp_connection::MCP_SANDBOX_STATE_META_CAPABILITY;
 use crate::mcp_connection::McpRuntimeEnvironment;
-use crate::mcp_connection::ToolFilter;
-use crate::mcp_connection::ToolInfo;
 use crate::mcp_connection::emit_duration;
-use crate::mcp_connection::filter_tools;
 use crate::mcp_connection::resolve_bearer_token;
-use crate::mcp_connection::tool_with_model_visible_input_schema;
 use crate::mcp_connection::validate_mcp_server_name;
+use crate::tools::ToolFilter;
+use crate::tools::ToolInfo;
+use crate::tools::filter_tools;
+use crate::tools::tool_with_model_visible_input_schema;
 use anyhow::Result;
 use anyhow::anyhow;
 use async_channel::Sender;
