+++ codex-rs/codex-mcp/src/lib.rs
@@ -2,7 +2,7 @@ pub use manager::McpConnectionManager;
 pub use mcp_connection::MCP_SANDBOX_STATE_META_CAPABILITY;
 pub use mcp_connection::McpRuntimeEnvironment;
 pub use mcp_connection::SandboxState;
-pub use mcp_connection::ToolInfo;
+pub use tools::ToolInfo;
 
 pub use mcp::CODEX_APPS_MCP_SERVER_NAME;
 pub use mcp::McpConfig;
@@ -36,7 +36,7 @@ pub use mcp::should_retry_without_scopes;
 pub use apps::filter_non_codex_apps_mcp_tools_only;
 pub use mcp::mcp_permission_prompt_is_auto_approved;
 pub use mcp::qualified_mcp_tool_name_prefix;
-pub use mcp_connection::declared_openai_file_input_param_names;
+pub use tools::declared_openai_file_input_param_names;
 
 pub(crate) mod apps;
 pub(crate) mod client;
@@ -44,4 +44,4 @@ pub(crate) mod elicitation;
 pub(crate) mod manager;
 pub(crate) mod mcp;
 pub(crate) mod mcp_connection;
-pub(crate) mod mcp_tool_names;
+pub(crate) mod tools;
