+++ codex-rs/codex-mcp/src/manager.rs
@@ -18,15 +18,15 @@ use crate::elicitation::ElicitationRequestManager;
 use crate::mcp::CODEX_APPS_MCP_SERVER_NAME;
 use crate::mcp::ToolPluginProvenance;
 use crate::mcp_connection::McpRuntimeEnvironment;
-use crate::mcp_connection::ToolInfo;
 use crate::mcp_connection::emit_duration;
 use crate::mcp_connection::emit_update;
-use crate::mcp_connection::filter_tools;
 use crate::mcp_connection::mcp_init_error_display;
-use crate::mcp_connection::qualify_tools;
 use crate::mcp_connection::startup_outcome_error_message;
-use crate::mcp_connection::tool_with_model_visible_input_schema;
 use crate::mcp_connection::transport_origin;
+use crate::tools::ToolInfo;
+use crate::tools::filter_tools;
+use crate::tools::qualify_tools;
+use crate::tools::tool_with_model_visible_input_schema;
 use anyhow::Context;
 use anyhow::Result;
 use anyhow::anyhow;
