+++ codex-rs/codex-mcp/src/mcp_connection_manager_tests.rs
@@ -11,11 +11,12 @@ use crate::client::elicitation_capability_for_server;
 use crate::declared_openai_file_input_param_names;
 use crate::elicitation::ElicitationRequestManager;
 use crate::elicitation::elicitation_is_rejected_by_policy;
-use crate::mcp_connection::ToolFilter;
-use crate::mcp_connection::ToolInfo;
-use crate::mcp_connection::filter_tools;
-use crate::mcp_connection::tool_with_model_visible_input_schema;
 use crate::mcp_connection::transport_origin;
+use crate::tools::ToolFilter;
+use crate::tools::ToolInfo;
+use crate::tools::filter_tools;
+use crate::tools::qualify_tools;
+use crate::tools::tool_with_model_visible_input_schema;
 use codex_config::Constrained;
 use codex_protocol::ToolName;
 use codex_protocol::protocol::GranularApprovalConfig;
